Add files via upload
This is the adding your money into the system screen. Right now the program adds data to a text file and if no number is put into the text box it will be put in as a zero in the file. Still have to work on the credit and debit buttons and also probably have something so it pops up an error if something other than numbers is put in.